Add files via upload
The ATM Simulator is a desktop application that simulates the functionalities of an Automated Teller Machine (ATM) through a graphical interface. It allows users to log in with their card number and PIN, check their account balance, withdraw cash, and deposit money. The system validates credentials, ensures sufficient funds for withdrawals, and provides error alerts for insufficient funds or incorrect inputs. The user-friendly interface includes navigable panels, buttons, menus, text fields, and display areas. Although it offers many features, it has limitations, such as simplified operations compared to real ATMs and a user interface that may not fully capture the responsiveness of a real ATM. Despite these limitations, the ATM Simulator is an effective tool for simulating banking operations, offering valuable insights and practical experience in a controlled, virtual environment.